VSL-80: Added util function for creating signatures
