+++ packages/client/src/components/SafeSettings.js
@@ -8,6 +8,7 @@ import {
   getProgressPercentageForSignersAmount,
   isAddr,
   formatAddress,
+  createSignature
 } from "../utils";
 
 const SignatureBar = ({ threshold, safeOwners }) => (
@@ -447,23 +448,8 @@ function SafeSettings({ address, web3, name, threshold, safeOwners }) {
     const thresholdToPersist = newThreshold ?? threshold;
 
     if (newOwner) {
-      const latestBlock = await web3.injectedProvider
-        .send([web3.injectedProvider.getBlock(true)])
-        .then(web3.injectedProvider.decode);
-
-      const { height, id } = latestBlock;
-      const newOwnerHex = Buffer.from(`Add account ${newOwner.address} as a signer.`).toString("hex");
-
-      const message = `${newOwnerHex}${id}`;
-      const messageHex = Buffer.from(message).toString("hex");
-
-      let sigResponse = await web3.injectedProvider
-        .currentUser()
-        .signUserMessage(messageHex);
-      const sigMessage =
-        sigResponse[0]?.signature?.signature ?? sigResponse[0]?.signature;
-      const keyIds = [sigResponse[0]?.keyId];
-      const signatures = [sigMessage];
+      const intent = `Add account ${newOwner.address} as a signer.`;
+      const { message, keyIds, signatures, height } = createSignature(web3, intent);
 
       await proposeAddSigner(
         formatAddress(newOwner.address),
@@ -475,24 +461,8 @@ function SafeSettings({ address, web3, name, threshold, safeOwners }) {
     }
 
     if (thresholdToPersist !== threshold) {
-
-      const latestBlock = await web3.injectedProvider
-        .send([web3.injectedProvider.getBlock(true)])
-        .then(web3.injectedProvider.decode);
-
-      const { height, id } = latestBlock;
-      const thresholdToPersistHex = Buffer.from(`Update the threshold of signers to ${thresholdToPersist}.`).toString("hex");
-
-      const message = `${thresholdToPersistHex}${id}`;
-      const messageHex = Buffer.from(message).toString("hex");
-
-      let sigResponse = await web3.injectedProvider
-        .currentUser()
-        .signUserMessage(messageHex);
-      const sigMessage =
-        sigResponse[0]?.signature?.signature ?? sigResponse[0]?.signature;
-      const keyIds = [sigResponse[0]?.keyId];
-      const signatures = [sigMessage];
+      const intent = `Update the threshold of signers to ${thresholdToPersist}.`;
+      const { message, keyIds, signatures, height } = createSignature(web3, intent);
 
       await updateThreshold(
         thresholdToPersist,
@@ -513,23 +483,8 @@ function SafeSettings({ address, web3, name, threshold, safeOwners }) {
 
   const onRemoveSafeOwnerSubmit = async (ownerToBeRemoved) => {
     if (ownerToBeRemoved) {
-      const latestBlock = await web3.injectedProvider
-        .send([web3.injectedProvider.getBlock(true)])
-        .then(web3.injectedProvider.decode);
-
-      const { height, id } = latestBlock;
-      const ownerToBeRemovedHex = Buffer.from(`Remove ${ownerToBeRemoved.address} as a signer.`).toString("hex");
-
-      const message = `${ownerToBeRemovedHex}${id}`;
-      const messageHex = Buffer.from(message).toString("hex");
-
-      let sigResponse = await web3.injectedProvider
-        .currentUser()
-        .signUserMessage(messageHex);
-      const sigMessage =
-        sigResponse[0]?.signature?.signature ?? sigResponse[0]?.signature;
-      const keyIds = [sigResponse[0]?.keyId];
-      const signatures = [sigMessage];
+      const intent = `Remove ${ownerToBeRemoved.address} as a signer.`;
+      const { message, keyIds, signatures, height } = createSignature(web3, intent);
 
       await proposeRemoveSigner(
         formatAddress(ownerToBeRemoved.address),
